change Int to Double (Volume)

+++ ProjectOOP/src/Process/Rule3.java
@@ -14,7 +14,7 @@ public class Rule3 extends Rules {
 
         //phương thức này dùng để tạo ra các đối tượng Rules và thêm nó vào trong ArrayList của lớp Rules
         for (int i=0;i<(data.size());i++){
-            int tmp= data.get(i).getVolume();
+            double tmp= data.get(i).getVolume();
             double temp=data.get(i).getVolume()*((data.get(i).getHigh())-(data.get(i).getLow()))/2*1000;
             arr.add(new Rules(i,tmp,temp));
         }
@@ -23,7 +23,7 @@ public class Rule3 extends Rules {
         Collections.sort(arr,new SortByVolume());
         for(int i=0;i<3;i++) {
             map.put("vol" + (i + 1), data.get(arr.get(i).getIndex()).getName());
-            map.put("v"+(i+1), Integer.toString(arr.get(i).getVolume()));
+            map.put("v"+(i+1), Double.toString(arr.get(i).getVolume()));
         }
 
         //sắp xếp theo VolumeValue và đưa các cặp key-value cần thiết vào map
