+++ ProjectOOP/src/Process/Rule9.java
@@ -45,7 +45,7 @@ public class Rule9 extends Rules {
         Collections.sort(array,new DataSortByVolumeValue());
         for(int i=0;i<array.size();i++) {
             map.put("TVname" + (i + 1),array.get(i).getName());
-            map.put("TVnum"+(i+1),Integer.toString(array.get(i).getVolume()));
+            map.put("TVnum"+(i+1),Double.toString(array.get(i).getVolume()));
         }
         map.put("totalT",Integer.toString(filterByName(data,"^THEP").getVolume()));
         for(int i=0;i<array.size();i++) {
